requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
Pillow>=8.0.0
timm>=0.5.4
numpy>=1.19.0
transformers>=4.37.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
pandas>=1.3.0
seaborn>=0.11.0
requests>=2.31.0
tqdm>=4.66.1

